requirements.txt
streamlit
pandas
pyspellchecker
openpyxl
